+++ application/src/main/java/nl/jixxed/eliteodysseymaterials/parser/messageprocessor/FSDJumpMessageProcessor.java
@@ -21,9 +21,9 @@ public class FSDJumpMessageProcessor implements MessageProcessor<FSDJump> {
         final String factionState = event.getSystemFaction().map(systemFaction -> systemFaction.getFactionState().orElse("")).orElse("");
 
         if (!starSystem.isBlank()) {
-            final double x = event.getStarPos().get(0);
-            final double y = event.getStarPos().get(1);
-            final double z = event.getStarPos().get(2);
+            final double x = event.getStarPos().get(0).doubleValue();
+            final double y = event.getStarPos().get(1).doubleValue();
+            final double z = event.getStarPos().get(2).doubleValue();
             EventService.publish(new FSDJumpJournalEvent(event, new StarSystem(starSystem, SystemEconomy.forKey(economy), SystemEconomy.forKey(secondEconomy), SystemGovernment.forKey(government), SystemSecurity.forKey(security), factionState, x, y, z), body));
         }
         EDDNService.fsdjump(event);
