+++ application/src/test/java/nl/jixxed/eliteodysseymaterials/service/PathServiceTest.java
@@ -19,6 +19,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Locale;
 
@@ -63,8 +64,8 @@ class PathServiceTest {
 
         EventService.publish(new LocationJournalEvent(new Location.LocationBuilder()
                 .withBody("A")
-                .withBodyID(1L)
-                .withSystemAddress(123L)
+                .withBodyID(BigInteger.valueOf(1L))
+                .withSystemAddress(BigInteger.valueOf(123L))
                 .withStarPos(List.of(Engineer.ODEN_GEIGER.getStarSystem().getX(),Engineer.ODEN_GEIGER.getStarSystem().getY(),Engineer.ODEN_GEIGER.getStarSystem().getZ()))
                 .withStarSystem(Engineer.ODEN_GEIGER.getStarSystem().getName())
                 .build(),Engineer.ODEN_GEIGER.getStarSystem(), "", "", true));
@@ -98,8 +99,8 @@ class PathServiceTest {
 
         EventService.publish(new LocationJournalEvent(new Location.LocationBuilder()
                 .withBody("A")
-                .withBodyID(1L)
-                .withSystemAddress(123L)
+                .withBodyID(BigInteger.valueOf(1L))
+                .withSystemAddress(BigInteger.valueOf(123L))
                 .withStarPos(List.of(Engineer.DOMINO_GREEN.getStarSystem().getX(),Engineer.DOMINO_GREEN.getStarSystem().getY(),Engineer.DOMINO_GREEN.getStarSystem().getZ()))
                 .withStarSystem(Engineer.DOMINO_GREEN.getStarSystem().getName())
                 .build(),Engineer.DOMINO_GREEN.getStarSystem(), "", "", true));
